orderingToolServer/src/main/java/ch/eschbach/cleanit/orderingToolServer/services/CustomerService.java
package ch.eschbach.cleanit.orderingToolServer.services;

import ch.eschbach.cleanit.orderingToolServer.model.Customer;
import ch.eschbach.cleanit.orderingToolServer.persistence.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
public class CustomerService {

    @Autowired
    private CustomerRepository customerRepo;

    public Customer getCustomerById(Long id) {
        return customerRepo.findById(id).orElse(null);
    }

    public List<Customer> getAllCustomers() {
        List<Customer> customers = customerRepo.findAll();
        return customers;
    }

    public void deleteCustomer(Customer customer) {
        if (customer == null) {
            throw new RuntimeException("customer parameter not set");
        }
        customerRepo.delete(customer);
    }
}
